refactor(catalog): extract component evidence panels
Keep URL ownership and current/historical selection in the coordinator so
presentation edits no longer require loading mutation context.
